Refactor BackFlow VMC functions into dedicated source
