+++ src/mVMC/makefile_src
@@ -26,6 +26,7 @@ backflow.c \
 backflow_measure.c \
 backflow_pf.c \
 backflow_slater.c \
+backflow_vmc.c \
 average.c \
 avevar.c \
 calgrn.c \
