+++ src/mVMC/vmcmake_real.c
@@ -42,54 +42,6 @@ which follows "The BSD 3-Clause License".
 #include "../pfupdates/pf_interface.h"
 #endif
 
-#ifdef MVMC_DEBUG_BF_REAL_UPDATE
-static void CheckBFRealUpdate(const char *label, const int *msa, const int *hopNum,
-                              const int *msaRef, const int *hopNumRef, const int rank) {
-  int qpidx, hop, idx;
-
-  for (qpidx = 0; qpidx < NQPFull; qpidx++) {
-    if (hopNum[qpidx] != hopNumRef[qpidx]) {
-      fprintf(stderr,
-              "error: BF real update mismatch (%s): rank=%d qp=%d hopNum=%d ref=%d\n",
-              label, rank, qpidx, hopNum[qpidx], hopNumRef[qpidx]);
-      MPI_Abort(MPI_COMM_WORLD, EXIT_FAILURE);
-    }
-    for (hop = 0; hop < hopNum[qpidx]; hop++) {
-      idx = qpidx*Nsize + hop;
-      if (msa[idx] != msaRef[idx]) {
-        fprintf(stderr,
-                "error: BF real update mismatch (%s): rank=%d qp=%d hop=%d msa=%d ref=%d\n",
-                label, rank, qpidx, hop, msa[idx], msaRef[idx]);
-        MPI_Abort(MPI_COMM_WORLD, EXIT_FAILURE);
-      }
-    }
-  }
-
-  for (idx = 0; idx < NQPFull*Nsite2*Nsite2; idx++) {
-    const double ref = creal(SlaterElmBF[idx]);
-    if (SlaterElmBF_real[idx] != ref) {
-      fprintf(stderr,
-              "error: BF real update mismatch (%s): rank=%d idx=%d slt=%.17e ref=%.17e\n",
-              label, rank, idx, SlaterElmBF_real[idx], ref);
-      MPI_Abort(MPI_COMM_WORLD, EXIT_FAILURE);
-    }
-  }
-}
-
-static void UpdateSlaterElmBF_real_checked(const char *label,
-                                           const int ma, const int ra, const int rb, const int u,
-                                           const int *eleCfg, const int *eleNum,
-                                           const int *eleProjBFCntOld, const int *eleProjBFCnt,
-                                           int *msa, int *hopNum,
-                                           int *msaRef, int *hopNumRef, const int rank) {
-  UpdateSlaterElmBF_fcmp(ma, ra, rb, u, eleCfg, eleNum, eleProjBFCnt, msaRef, hopNumRef,
-                         SlaterElmBF);
-  UpdateSlaterElmBF_real(ma, ra, rb, u, eleCfg, eleNum, eleProjBFCntOld, eleProjBFCnt, msa, hopNum,
-                         SlaterElmBF_real);
-  CheckBFRealUpdate(label, msa, hopNum, msaRef, hopNumRef, rank);
-}
-#endif
-
 void VMCMakeSample_real(MPI_Comm comm) {
   int outStep, nOutStep;
   int inStep, nInStep;
@@ -529,321 +481,5 @@ int makeInitialSample_real(int *eleIdx, int *eleCfg, int *eleNum, int *eleProjCn
   return 0;
 }
 
-void VMC_BF_MakeSample_real(MPI_Comm comm) {
-  int outStep, nOutStep;
-  int inStep, nInStep;
-  UpdateType updateType;
-  int mi, mj, ri, rj, s, t, i;
-  int nAccept = 0;
-  int sample;
-  int tmp_i;
-
-  double logIpOld, logIpNew; /* logarithm of inner product <phi|L|x> */ // is this ok ? TBC
-  int projCntNew[NProj];
-  int projBFCntNew[16 * Nsite * Nrange]; // For BackFlow
-  int msaTmp[NQPFull * Nsize], icount[NQPFull]; // For BackFlow
-#ifdef MVMC_DEBUG_BF_REAL_UPDATE
-  int msaRef[NQPFull * Nsize], icountRef[NQPFull];
-#endif
-  double pfMNew_real[NQPFull];
-  double x, w; // TBC x will be complex number
-
-  int qpStart, qpEnd;
-  int rejectFlag;
-  int rank, size;
-  MPI_Comm_size(comm, &size);
-  MPI_Comm_rank(comm, &rank);
-
-  SplitLoop(&qpStart, &qpEnd, NQPFull, rank, size);
-
-  StartTimer(30);
-  if (BurnFlag == 0) {
-    makeInitialSampleBF_real(TmpEleIdx, TmpEleCfg, TmpEleNum, TmpEleProjCnt, TmpEleProjBFCnt,
-                             qpStart, qpEnd, comm);
-    //makeInitialSample(TmpEleIdx,TmpEleCfg,TmpEleNum,TmpEleProjCnt,
-    //                  qpStart,qpEnd,comm);
-  } else {
-    //copyFromBurnSample(TmpEleIdx,TmpEleCfg,TmpEleNum,TmpEleProjCnt);
-    copyFromBurnSampleBF(TmpEleIdx);
-    MakeSlaterElmBF_fcmp(TmpEleNum, TmpEleProjBFCnt);
-#pragma omp parallel for default(shared) private(tmp_i)
-    for(tmp_i=0;tmp_i<NQPFull*(2*Nsite)*(2*Nsite);tmp_i++) SlaterElmBF_real[tmp_i]= creal(SlaterElmBF[tmp_i]);
-  }
-
-  CalculateMAll_BF_real(TmpEleIdx, qpStart, qpEnd);
-  // printf("DEBUG: maker1: PfM=%lf\n",creal(PfM[0]));
-  logIpOld = CalculateLogIP_real(PfM_real, qpStart, qpEnd, comm);
-  if (!isfinite(logIpOld)) {
-    if (rank == 0) fprintf(stderr, "waring: VMCMakeSample remakeSample logIpOld=%e\n", creal(logIpOld)); //TBC
-    //    makeInitialSample(TmpEleIdx,TmpEleCfg,TmpEleNum,TmpEleProjCnt,
-    //                    qpStart,qpEnd,comm);
-    makeInitialSampleBF_real(TmpEleIdx, TmpEleCfg, TmpEleNum, TmpEleProjCnt, TmpEleProjBFCnt,
-                             qpStart, qpEnd, comm);
-
-    CalculateMAll_BF_real(TmpEleIdx, qpStart, qpEnd);
-    //printf("DEBUG: maker2: PfM=%lf\n",creal(PfM[0]));
-    logIpOld = CalculateLogIP_real(PfM_real, qpStart, qpEnd, comm);
-    BurnFlag = 0;
-  }
-  StopTimer(30);
-
-  nOutStep = (BurnFlag == 0) ? NVMCWarmUp + NVMCSample : NVMCSample + 1;
-  nInStep = NVMCInterval * Nsite;
-
-  for (i = 0; i < Counter_max; i++) Counter[i] = 0;  /* reset counter */
-
-  for (outStep = 0; outStep < nOutStep; outStep++) {
-    for (inStep = 0; inStep < nInStep; inStep++) {
-
-      updateType = getUpdateType(NExUpdatePath);
-
-      if (updateType == HOPPING) { /* hopping */
-        Counter[0]++;
-        AddBFProfileCounter(BFPROF_HOP_TRY, 1);
-
-        StartTimer(31);
-        makeCandidate_hopping(&mi, &ri, &rj, &s, &rejectFlag,
-                              TmpEleIdx, TmpEleCfg);
-        StopTimer(31);
-
-        if (rejectFlag) {
-          AddBFProfileCounter(BFPROF_HOP_CANDIDATE_REJECT, 1);
-          continue;
-        }
-        AddBFProfileCounter(BFPROF_HOP_VALID, 1);
-
-        StartTimer(32);
-        StartTimer(60);
-        /* The mi-th electron with spin s hops to site rj */
-        updateEleConfig(mi, ri, rj, s, TmpEleIdx, TmpEleCfg, TmpEleNum);
-        UpdateProjCnt(ri, rj, s, projCntNew, TmpEleProjCnt, TmpEleNum);
-        MakeProjBFCnt(projBFCntNew, TmpEleNum);
-        StopTimer(60);
-        StartTimer(64);
-#ifdef MVMC_DEBUG_BF_REAL_UPDATE
-        UpdateSlaterElmBF_real_checked("hopping proposal", mi, ri, rj, s, TmpEleCfg, TmpEleNum,
-                                       TmpEleProjBFCnt, projBFCntNew, msaTmp, icount, msaRef, icountRef, rank);
-#else
-        /* Real BF sampling updates SlaterElmBF_real directly; SlaterElmBF is not kept current on this hot path. */
-        UpdateSlaterElmBF_real(mi, ri, rj, s, TmpEleCfg, TmpEleNum, TmpEleProjBFCnt, projBFCntNew, msaTmp, icount,
-                               SlaterElmBF_real);
-#endif
-        StopTimer(64);
-
-        StartTimer(61);
-        //CalculateNewPfM2(mi,s,pfMNew,TmpEleIdx,qpStart,qpEnd);
-        //CalculateNewPfM2_real(mi,s,pfMNew_real,TmpEleIdx,qpStart,qpEnd);
-        CalculateNewPfMBF_real(icount, msaTmp, pfMNew_real, TmpEleIdx, qpStart, qpEnd, SlaterElmBF_real);
-
-        //printf("DEBUG: out %d in %d pfMNew=%lf \n",outStep,inStep,creal(pfMNew[0]));
-        StopTimer(61);
-
-        StartTimer(62);
-        /* calculate inner product <phi|L|x> */
-        //logIpNew = CalculateLogIP_fcmp(pfMNew,qpStart,qpEnd,comm);
-        logIpNew = CalculateLogIP_real(pfMNew_real, qpStart, qpEnd, comm);
-        StopTimer(62);
-
-        /* Metroplis */
-        x = LogProjRatio(projCntNew, TmpEleProjCnt);
-        w = exp(2.0 * (x + (logIpNew - logIpOld)));
-        if (!isfinite(w)) w = -1.0; /* should be rejected */
-
-        if (w > genrand_real2()) { /* accept */
-          // UpdateMAll will change SlaterElm, InvM (including PfM)
-          StartTimer(63);
-          UpdateMAll_BF_real(icount, msaTmp, PfM_real, TmpEleIdx, qpStart, qpEnd);
-          //          UpdateMAll_real(mi,s,TmpEleIdx,qpStart,qpEnd);
-          //            UpdateMAll(mi,s,TmpEleIdx,qpStart,qpEnd);
-          StopTimer(63);
-
-          for (i = 0; i < NProj; i++) TmpEleProjCnt[i] = projCntNew[i];
-          for (i = 0; i < 16 * Nsite * Nrange; i++) TmpEleProjBFCnt[i] = projBFCntNew[i];
-          logIpOld = logIpNew;
-          nAccept++;
-          Counter[1]++;
-          AddBFProfileCounter(BFPROF_HOP_ACCEPT, 1);
-        } else { /* reject */
-          AddBFProfileCounter(BFPROF_HOP_METROPOLIS_REJECT, 1);
-          revertEleConfig(mi, ri, rj, s, TmpEleIdx, TmpEleCfg, TmpEleNum);
-          StartTimer(64);
-#ifdef MVMC_DEBUG_BF_REAL_UPDATE
-          UpdateSlaterElmBF_real_checked("hopping reject", mi, rj, ri, s, TmpEleCfg, TmpEleNum,
-                                         projBFCntNew, TmpEleProjBFCnt, msaTmp, icount, msaRef, icountRef, rank);
-#else
-          UpdateSlaterElmBF_real(mi, rj, ri, s, TmpEleCfg, TmpEleNum, projBFCntNew, TmpEleProjBFCnt, msaTmp, icount,
-                                 SlaterElmBF_real);
-#endif
-          StopTimer(64);
-        }
-        StopTimer(32);
-
-      } else if (updateType == EXCHANGE) { /* exchange */
-        Counter[2]++;
-        AddBFProfileCounter(BFPROF_EXCHANGE_TRY, 1);
-
-        StartTimer(31);
-        makeCandidate_exchange(&mi, &ri, &rj, &s, &rejectFlag,
-                               TmpEleIdx, TmpEleCfg, TmpEleNum);
-        StopTimer(31);
-
-        if (rejectFlag) {
-          AddBFProfileCounter(BFPROF_EXCHANGE_CANDIDATE_REJECT, 1);
-          continue;
-        }
-        AddBFProfileCounter(BFPROF_EXCHANGE_VALID, 1);
-
-        StartTimer(33);
-        StartTimer(65);
-
-        /* The mi-th electron with spin s exchanges with the electron on site rj with spin 1-s */
-        t = 1 - s;
-        mj = TmpEleCfg[rj + t * Nsite];
-
-        /* The mi-th electron with spin s hops to rj */
-        updateEleConfig(mi, ri, rj, s, TmpEleIdx, TmpEleCfg, TmpEleNum);
-        UpdateProjCnt(ri, rj, s, projCntNew, TmpEleProjCnt, TmpEleNum);
-        /* The mj-th electron with spin t hops to ri */
-        updateEleConfig(mj, rj, ri, t, TmpEleIdx, TmpEleCfg, TmpEleNum);
-        UpdateProjCnt(rj, ri, t, projCntNew, projCntNew, TmpEleNum);
-
-        StopTimer(65);
-        StartTimer(66);
-
-        CalculateNewPfMTwo2_real(mi, s, mj, t, pfMNew_real, TmpEleIdx, qpStart, qpEnd);
-        StopTimer(66);
-        StartTimer(67);
-
-        /* calculate inner product <phi|L|x> */
-        logIpNew = CalculateLogIP_real(pfMNew_real, qpStart, qpEnd, comm);
-
-        StopTimer(67);
-
-        /* Metroplis */
-        x = LogProjRatio(projCntNew, TmpEleProjCnt);
-        w = exp(2.0 * (x + (logIpNew - logIpOld))); //TBC
-        if (!isfinite(w)) w = -1.0; /* should be rejected */
-
-        if (w > genrand_real2()) { /* accept */
-          StartTimer(68);
-          UpdateMAllTwo_real(mi, s, mj, t, ri, rj, TmpEleIdx, qpStart, qpEnd);
-          StopTimer(68);
-
-          for (i = 0; i < NProj; i++) TmpEleProjCnt[i] = projCntNew[i];
-          logIpOld = logIpNew;
-          nAccept++;
-          Counter[3]++;
-          AddBFProfileCounter(BFPROF_EXCHANGE_ACCEPT, 1);
-        } else { /* reject */
-          AddBFProfileCounter(BFPROF_EXCHANGE_METROPOLIS_REJECT, 1);
-          revertEleConfig(mj, rj, ri, t, TmpEleIdx, TmpEleCfg, TmpEleNum);
-          revertEleConfig(mi, ri, rj, s, TmpEleIdx, TmpEleCfg, TmpEleNum);
-        }
-        StopTimer(33);
-      }
-
-      if (nAccept > Nsite) {
-        StartTimer(34);
-        /* recal PfM and InvM */
-        //CalculateMAll_real(TmpEleIdx,qpStart,qpEnd);
-        //printf("DEBUG: maker3: PfM=%lf\n",creal(PfM[0]));
-        CalculateMAll_BF_real(TmpEleIdx, qpStart, qpEnd);
-        logIpOld = CalculateLogIP_real(PfM_real, qpStart, qpEnd, comm);
-        StopTimer(34);
-        nAccept = 0;
-      }
-    } /* end of instep */
-
-    StartTimer(35);
-    /* save Electron Configuration */
-    if (outStep >= nOutStep - NVMCSample) {
-      sample = outStep - (nOutStep - NVMCSample);
-      saveEleConfigBF(sample, logIpOld, TmpEleIdx, TmpEleCfg, TmpEleNum, TmpEleProjCnt, TmpEleProjBFCnt);
-      //      saveEleConfig(sample,logIpOld,TmpEleIdx,TmpEleCfg,TmpEleNum,TmpEleProjCnt);
-    }
-    StopTimer(35);
-
-  } /* end of outstep */
-
-  //copyToBurnSample(TmpEleIdx,TmpEleCfg,TmpEleNum,TmpEleProjCnt);
-  copyToBurnSampleBF(TmpEleIdx);
-  BurnFlag = 1;
-  return;
-}
-
-int makeInitialSampleBF_real(int *eleIdx, int *eleCfg, int *eleNum, int *eleProjCnt, int *eleProjBFCnt,
-                             const int qpStart, const int qpEnd, MPI_Comm comm) {
-  const int nsize = Nsize;
-  const int nsite2 = Nsite2;
-  int flag = 1, flagRdc, loop = 0;
-  int ri, mi, si, msi, rsi;
-  int rank, size;
-  int tmp_i;
-
-  MPI_Comm_size(comm, &size);
-  MPI_Comm_rank(comm, &rank);
-
-  do {
-    /* initialize */
-#pragma omp parallel for default(shared) private(msi)
-    for (msi = 0; msi < nsize; msi++) eleIdx[msi] = -1;
-#pragma omp parallel for default(shared) private(rsi)
-    for (rsi = 0; rsi < nsite2; rsi++) eleCfg[rsi] = -1;
-
-    /* local spin */
-    for (ri = 0; ri < Nsite; ri++) {
-      if (LocSpn[ri] == 1) {
-        do {
-          mi = gen_rand32() % Ne;
-          si = (genrand_real2() < 0.5) ? 0 : 1;
-        } while (eleIdx[mi + si * Ne] != -1);
-        eleCfg[ri + si * Nsite] = mi;
-        eleIdx[mi + si * Ne] = ri;
-      }
-    }
-
-    /* itinerant electron */
-    for (si = 0; si < 2; si++) {
-      for (mi = 0; mi < Ne; mi++) {
-        if (eleIdx[mi + si * Ne] == -1) {
-          do {
-            ri = gen_rand32() % Nsite;
-          } while (eleCfg[ri + si * Nsite] != -1 || LocSpn[ri] == 1);
-          eleCfg[ri + si * Nsite] = mi;
-          eleIdx[mi + si * Ne] = ri;
-        }
-      }
-    }
-
-    /* EleNum */
-#pragma omp parallel for default(shared) private(rsi)
-#pragma loop noalias
-    for (rsi = 0; rsi < nsite2; rsi++) {
-      eleNum[rsi] = (eleCfg[rsi] < 0) ? 0 : 1;
-    }
-
-    MakeProjCnt(eleProjCnt, eleNum);
-    MakeProjBFCnt(eleProjBFCnt, eleNum);
-
-    MakeSlaterElmBF_fcmp(eleNum, eleProjBFCnt);
-#pragma omp parallel for default(shared) private(tmp_i)
-    for(tmp_i=0;tmp_i<NQPFull*(2*Nsite)*(2*Nsite);tmp_i++) SlaterElmBF_real[tmp_i]= creal(SlaterElmBF[tmp_i]);
-
-    flag = CalculateMAll_BF_real(eleIdx, qpStart, qpEnd);
-    //printf("DEBUG: maker4: PfM=%lf\n",creal(PfM[0]));
-    if (size > 1) {
-      MPI_Allreduce(&flag, &flagRdc, 1, MPI_INT, MPI_MAX, comm);
-      flag = flagRdc;
-    }
-
-    loop++;
-    if (loop > 100) {
-      if (rank == 0) fprintf(stderr, "error: makeInitialSample: Too many loops\n");
-      MPI_Abort(MPI_COMM_WORLD, EXIT_FAILURE);
-    }
-  } while (flag > 0);
-
-  return 0;
-}
 
 #endif
